+++ api/v1/handler/user.go
@@ -1,0 +1,29 @@
+package handler
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/labstack/echo"
+)
+
+func UserID(c echo.Context) (err error) {
+	var userid string = c.Param("id")
+	name := GetName(userid)
+	return c.String(http.StatusOK, "Hello, World!"+name)
+}
+
+func GetName(id string) string {
+	switch id {
+	case "1":
+		fmt.Println("aaa")
+		return "aaa"
+	case "2":
+		fmt.Println("bbb")
+		return "bbb"
+	default:
+		fmt.Println("not found")
+		return "not found"
+	}
+	// return ""
+}
