+++ api/v1/routes.go
@@ -1,0 +1,12 @@
+package v1
+
+import (
+	"github.com/labstack/echo"
+	"github.com/nakapon9517/sample/api/v1/handler"
+)
+
+// _SetupRoutes :
+// ルート設定
+func SetupRoutes(ver *echo.Group) {
+	ver.GET("/users/:id", handler.UserID)
+}
