+++ go.mod
@@ -1,5 +1,14 @@
-module example.com/go-mod-test
+module github.com/nakapon9517/sample
 
-go 1.14
+go 1.13
 
-require github.com/labstack/echo/v4 v4.1.16
+require (
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
+	github.com/labstack/echo v3.3.10+incompatible
+	github.com/labstack/gommon v0.3.0 // indirect
+	github.com/mattn/go-colorable v0.1.6 // indirect
+	github.com/valyala/fasttemplate v1.1.0 // indirect
+	golang.org/x/crypto v0.0.0-20200221231518-2aa609cf4a9d // indirect
+	golang.org/x/net v0.0.0-20200226121028-0de0cce0169b // indirect
+	golang.org/x/text v0.3.2 // indirect
+)
